requirements.txt
stravalib==2.4.0
garminconnect==0.2.28
requests>=2.31.0
schedule>=1.2.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0